% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.ML_AIC.R
\name{Optimal.KDE_AIC}
\alias{Optimal.KDE_AIC}
\title{To Find Optimal Band Width Using AIC}
\usage{
Optimal.KDE_AIC(data, mplot = FALSE, mkernel = "gaussian")
}
\arguments{
\item{data}{a numeric vector}

\item{mplot}{by default `FALSE`, if TRUE then returns the AIC vs. h plot}

\item{mkernel}{by default `"gaussian"`. It can be any of the kernel supported by `density` function}
}
\value{
Optimal bandwidth with scale adjustment
}
\description{
To Find Optimal Band Width Using AIC
}
\examples{
install.packages("nor1Mix")
library(nor1mix) # Need to install
ex2 <- norMix(mu = c(-2, 0,2),sigma = c(0.6,0.4,0.6),w = c(0.3,0.4,0.3))
data2 <- rnorMix(300,ex2)  #generated data from Mixture Normal
h2 <- Optimal.KDE_AIC(data2)
plot(density(data2,h2),col = "cornflowerblue")
}
