% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.ML_AIC.R
\name{Optimal.KDE_ML}
\alias{Optimal.KDE_ML}
\title{To Find Optimal Band Width Using K-Fold Maximum Likelihood Method}
\usage{
Optimal.KDE_ML(data, k = 2, mplot = FALSE, mkernel = "gaussian")
}
\arguments{
\item{data}{a numeric vector}

\item{k}{an integer greater than 1, by default `k = 2`}

\item{mplot}{by default `FALSE`, if TRUE then returns the mean log-likelihood vs. h plot}

\item{mkernel}{by default `"gaussian"`. It can be any of the kernel supported by `density` function}
}
\value{
Optimal bandwidth with scale adjustment
}
\description{
To Find Optimal Band Width Using K-Fold Maximum Likelihood Method
}
\examples{
install.packages("nor1mix")
library(nor1mix)  #need to install
ex1 <- norMix(mu = c(-2, 2),sigma = c(1, 0.8),w = c(.4, .6))
data1 <- rnorMix(300,ex1)  #generated data from Mixture Normal
h1 <- Optimal.KDE_ML(data1)
plot(density(data1,h1),col = "cornflowerblue")
}
